+++ changelog.d/fixed-dashboard-vpn-active-detection-now-matches-5f3a.md
@@ -1,0 +1,9 @@
+_2026-04-26_
+
+## English
+
+Dashboard 'VPN active' detection now matches the kmod/zygisk filter for renamed interfaces (if<N> from issue #86, MyVPN, wg-client, ...).
+
+## Русский
+
+Определение 'VPN активен' на дашборде теперь совпадает с правилами фильтра kmod/zygisk для переименованных интерфейсов (if<N> из issue #86, MyVPN, wg-client и т. п.).
